+++ lib/app_module.dart
@@ -2,9 +2,9 @@ import 'package:flutter_modular/flutter_modular.dart';
 
 import 'features/authentication/binds/authentication_module.dart';
 import 'features/authentication/presentation/controllers/session_controller.dart';
+import 'features/book/binds/book_module.dart';
 import 'features/home/binds/home_module.dart';
 import 'features/onboarding/binds/onboarding_module.dart';
-import 'routes/home_routes.dart';
 import 'routes/routes.dart';
 
 class AppModule extends Module {
@@ -19,6 +19,7 @@ class AppModule extends Module {
   List<Module> get imports => [
         AuthenticationModule(),
         HomeModule(),
+        BookModule(),
       ];
 
   @override
@@ -34,6 +35,10 @@ class AppModule extends Module {
         ModuleRoute(
           HomeRoutes.homeModuleRoute,
           module: HomeModule(),
+        ),
+        ModuleRoute(
+          BookRoutes.bookModuleRoute,
+          module: BookModule(),
         ),
       ];
 }