+++ lib/features/book/presentation/controllers/book_details_controller.g.dart
@@ -1,0 +1,26 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'book_details_controller.dart';
+
+// **************************************************************************
+// StoreGenerator
+// **************************************************************************
+
+// ignore_for_file: non_constant_identifier_names, unnecessary_brace_in_string_interps, unnecessary_lambdas, prefer_expression_function_bodies, lines_longer_than_80_chars, avoid_as, avoid_annotating_with_dynamic, no_leading_underscores_for_local_identifiers
+
+mixin _$BookDetailsController on _BookDetailsController, Store {
+  Computed<bool>? _$isValidFormComputed;
+
+  @override
+  bool get isValidForm =>
+      (_$isValidFormComputed ??= Computed<bool>(() => super.isValidForm,
+              name: '_BookDetailsController.isValidForm'))
+          .value;
+
+  @override
+  String toString() {
+    return '''
+isValidForm: ${isValidForm}
+    ''';
+  }
+}