+++ lib/features/book/presentation/widgets/star_widget.dart
@@ -1,0 +1,14 @@
+import 'package:flutter/material.dart';
+
+class StarWidget extends StatelessWidget {
+  const StarWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return const Icon(
+      Icons.star_outlined,
+      color: Colors.yellow,
+      size: 25,
+    );
+  }
+}